+++ .travis.yml
@@ -1,4 +1,4 @@
-sudo: false
+s	sesudo: false
 language: python
 python:
   - "3.4"
